feat: load paginated results in dashboard
